requirements.txt
fuzzywuzzy
fastapi
uvicorn
pypdf
langchain
openai
python-dotenv
psycopg2-binary
python-multipart
pymilvus
pyyaml
python-Levenshtein
numpy
pandas
pypdf2
groq
einops
sqlalchemy
pydantic-settings